requirements.txt
streamlit>=1.33
groq>=0.37
python-dotenv>=1.0
pandas>=2.1
openai>=1.30.0
httpx>=0.27.0

# Report generation
python-docx>=1.1.0
reportlab>=4.0

# Visualization
matplotlib>=3.8

